Move the Task content to their own components
